+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **breaking:** moved `sample` to `Noise` trait
 - **breaking:** removed `SampleWithSeed` trait
 - **breaking:** removed `Weighted`'s `seed` and `frequency` methods in favor of  the ones provided by `Noise`
+- **breaking:** `frequency` can now be called with a noise argument
 - **added:** `add_seed` modifier
 - **added:** `translate_*` modifiers
 - **added:** `add`, `sub`, `mul`, `div`, `rem`, `min`, `max`, `clamp`, `pow`, `neg`, `abs`, `floor`, `ceil`, `round`, `lerp` modifiers